+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 
 Unreleased / master
 -------------------
+* The ISPyB database ORM model version can now be interrogated via ``ispyb.sqlalchemy.__schema_version__``.
 
 6.10.0 (2022-01-13)
 -------------------